Revert "solving warnings"
This reverts commit d5dfbdbb0f6b0b40474adcc023152b397907eb19.

+++ include/Order.h
@@ -26,7 +26,7 @@ public:
 
 
     //setters
-    void setTips(float t){tips=t;}
+    void setTips(unsigned int t){tips=t;}
     void setDifficulty(Dif d){difficulty=d;}
     void setMaxDuration(unsigned int seconds){maxDuration=std::chrono::seconds(seconds);}
